[FIX] Property account selection during tests
By adding revenue accounts and property defaults in demo data as well as
installing a chart of accounts when running tests, there are actually two
property defaults for the revenue and income accounts. One of each (those
from the chart of accounts) have a default tax set, which can cause a
difference in the total amounts of the source invoice and the destination
invoice preventing the destination invoice from being confirmed, failing
the tests.