[11.0][FIX] partner_multi_company: make display_name stored
This is needed because after installation this field becomes
unsearchable and unsortable. Which is not explicitly changed in
this module and as such can be considered an undesired yield.

it is probably not a fix for the root cause but it is at
at least a fix for a long-standing problem