[FIX] product_multi_company: integrate tests with product_variant_configurator
When this module is integrated with `product_variant_configurator`, tests fail because the latter makes some black magic on the `name` field.

I switch the tests to rely on the `description_sale` field instead, which isn't touched by that module, and is also a field delegated from the related `product.template`.

BTW it doesn't make sense to modify provided permissions on a test that checks access errors, so instead I switched them to use users in a group that should have access in production.

@Tecnativa TT27288