[FIX] partner_multi_company: reconciliation compatibility
This line from the reconciliation widget
https://github.com/odoo/odoo/blob/12.0/addons/account/models/reconciliation_widget.py#L103
substitutes any occurrence of res_partner that could come in the
where_clause coming from ir_rules with the `p3` join alias of the
subsequent query.

In the case we've got a subselect, that alias won't make sense. So we're
preventing that situation renaming the table and the affected column
where such substitution was being made.

TT30346