[15.0][IMP] purchase_sale_inter_company:
If purchase_exception or sale_exception modules are installed
we have an issue in tests, because order's confirm methods won't be called.
Possible fix is to set ignore_exception=True on orders in tests.
Exception check could be skipped by setting check_exception=False,
as mentioned in base_exception, but in fact it doesn't work.