[FIX] account_invoice_inter_company: context can cause a UserError
Depending on the path used to reach the invoice, there can be a
default_journal_id in the context when the Form object used to create
the corresponding invoice on the other company is called. The journal id
is typically of the wrong type, which causes a UserError exception to be
raised, because the type of the journal and the type of the mirror
invoice don't match.

This commit fixes the issue by forcing the journal in the
default_journal_id context key in the call the Form