[FIX] purchase_sale_inter_company: fix demo data
Just changing these fields to make it easier to test the module.

Technically a bug in "stock" module's demo data. Contact Jeff Lawson is
also the "address" of warehouse "Chicago 1".
The fields property_stock_customer and supplier are set to
"Inter-warehouse transit" on Jeff Lawson for company San Francisco
and not for company Chicago as it should be.

Since there are no routes set for location "Inter-warehouse transit",
creating an inter-company PO from Chicago to SF fails;
when the inter-company SO is automatically created, with delivery
address towards the Chicago Warehouse i.e. Jeff Lawson, the creation
of the SO's picking fails as there is no route available.