[IMP-FIX] sale_product_company: define table in m2m model
When defining an m2m template it is convenient to define a unique named table to avoid conflicts with fields from other modules. This module is conflicting with the field of the abstract model ‘multi.company.abstract’ of the module base_multi_company which is extended in the model ‘product.template’ in the module product_multi_company as the same table is defined. To avoid conflicting with this abstract model, the m2m field of this module is renamed.

TT48972