[FIX] sale_product_company: fix test_sale_partner_company
Make tests compatible with other modules that hide the company_id field. For example, product_multi_company hides the company_id field in the form and using Form to run the tests causes a conflict. To solve this, the functionality of the form is replicated directly in the tests without using Form.

TT48972