[IMP] make picking state syncing configurable
By default picking state is forced to Done when SO picking is done, but if picking
sync fails, eg if there is a mismatch in SO and PO move lines, user can't correct
anything since the picking is already set to Done.

This may be acceptable in cases where 100% of the syncs go OK but we make the
setting configurable to support situations where manual corrections are needed.