[IMP] purchase_sale_stock_inter_company: Enhance validation logic Create two dedicated functions to determine when it is an intercompany reception or an intercompany delivery, and use the same logic in multiple places.
We used the locations to cover the dropshipping case and deliveries/receptions in multiples steps, and the respective many2one field to know if it is a purchase or sale intercompany transaction.
We do not use the partner because a picking can be created manually using this contact, but the specific fields for intercompany transactions are the many2one fields.