[FIX] partner_multi_company: Keep partner companies aligned
When a new company is created with user_ids, the update of the users in done only on cache level so the user's partner's company is not aligned.
Propagate user default company
Write on users triggers a write on the partners that validates the partners' companies, but partners' companies haven't been aligned yet so it raises the partner's companies constraint.
Align the companies in the partners first, so that writing on users (and therefore on partners) will satisfy the constraint.